+++ PROJECT-Rogue/Assets/Scripts/Weapons/ProjectileRiffle.cs
@@ -10,5 +10,5 @@ public class ProjectileRiffle : Projectile
         Vector2 ownerVelocity = (whoAttacks.GetComponent<Rigidbody2D>().velocity / 5);
         obj.GetComponent<Bullet>().SetParameters(whoAttacks, damage + whoAttacks.Damage, bulletSpeed, 1, ownerVelocity);
     }
-    public ProjectileRiffle(float attackSpeed = 10, float damage = 15, float bulletSpeed = 20, float bulletSize = 1) : base(attackSpeed, damage, bulletSpeed, bulletSize) { }
+    public ProjectileRiffle(float attackSpeed = 10, int damage = 15, float bulletSpeed = 20, float bulletSize = 1) : base(attackSpeed, damage, bulletSpeed, bulletSize) { }
 }
